upload via email/password, search projects by search term, various fixes
