--- ExampleCode/C++/GET/makefile
@@ -1,5 +1,0 @@
-CC		= g++
-CFLAGS = -O0 -g -Wall -Wextra -pedantic-errors  -w
-
-all:
-	$(CC) $(CFLAGS) GET_curl.cpp -o get.out -lcurl
