+++ ExampleCode/C++/PreAPI/README.md
@@ -1,0 +1,7 @@
+GET
+===
+
+These programs will include examples of getting data off of iSENSE/rSENSE into a C/C++ program.
+
+Currently, I have used the libcurl library with success to get the HTML of a page.
+It should also be possible to get JSON, and to then parse it using parson or another JSON C/C++ library.
